Added delete and rescinfo RPC

+++ server/main/src/com/smartnote/server/resource/FileResource.java
@@ -2,12 +2,15 @@ package com.smartnote.server.resource;
 
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.file.Files;
 import java.util.Objects;
+
+import com.smartnote.server.util.FileUtils;
 
 /**
  * <p>
@@ -29,13 +32,15 @@ class FileResource implements Resource {
      */
     FileResource(File file, AccessMode mode) {
         this.file = Objects.requireNonNull(file, "file must not be null");
-        this.mode = mode;
+        this.mode = Objects.requireNonNull(mode, "mode must not be null");
     }
 
     @Override
     public InputStream openInputStream() throws SecurityException, IOException {
         if (!mode.hasRead())
             throw new SecurityException("No read permission");
+        if (!file.exists())
+            throw new FileNotFoundException("File does not exist");
         return new FileInputStream(file);
     }
 
@@ -56,6 +61,22 @@ class FileResource implements Resource {
     public void delete() throws SecurityException, IOException {
         if (!mode.hasDelete())
             throw new SecurityException("No delete permission");
+        if (!file.exists())
+            throw new FileNotFoundException("File does not exist");
         file.delete();
+    }
+
+    @Override
+    public long size() throws SecurityException, IOException {
+        if (!mode.hasRead())
+            throw new SecurityException("No read permission");
+
+        if (!file.exists())
+            throw new FileNotFoundException("File does not exist");
+
+        if (file.isDirectory())
+            return FileUtils.getDirectorySize(file);
+
+        return file.length();
     }
 }
